refactor: change validate user service

+++ src/main/java/ru/artq/reminders/api/service/UserService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.artq.reminders.api.dto.UserDto;
+import ru.artq.reminders.api.exception.NotFoundException;
 import ru.artq.reminders.api.service.helper.ServiceHelper;
 import ru.artq.reminders.api.util.ConverterDto;
 import ru.artq.reminders.store.entity.UserEntity;
@@ -36,16 +37,25 @@ public class UserService {
     @Transactional
     public UserDto updateUser(Long userId, String username, String email, String password) {
         UserEntity user = serviceHelper.findUserById(userId);
-        serviceHelper.checkUsernameExist(username);
-        serviceHelper.checkEmailExist(email);
-        user.setUsername(username);
-        user.setEmail(email);
-        user.setPassword(password);
+        if (username != null && !username.isBlank()) {
+            serviceHelper.checkUsernameExist(username);
+            user.setUsername(username);
+        }
+        if (email != null && !email.isBlank() && !email.equals(user.getEmail())) {
+            serviceHelper.checkEmailExist(email);
+            user.setEmail(email);
+        }
+        if (password != null && !password.isBlank()) {
+            user.setPassword(password);
+        }
         return ConverterDto.userEntityToDto(userRepository.save(user));
     }
 
     @Transactional
     public void deleteUser(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new NotFoundException("User with ID '%d' not found.".formatted(userId));
+        }
         userRepository.deleteById(userId);
     }
 }
